refactor: move ecstore rebalance support helpers
